Official release v11.0.0

+++ Projects/NUCLEO-L152RE/Applications/IKS4A1/GyroscopeCalibration/Inc/motion_gc_manager.h
@@ -6,7 +6,7 @@
   ******************************************************************************
   * @attention
   *
-  * Copyright (c) 2023 STMicroelectronics.
+  * Copyright (c) 2024 STMicroelectronics.
   * All rights reserved.
   *
   * This software is licensed under terms that can be found in the LICENSE file
@@ -36,13 +36,13 @@ extern "C" {
 /* Imported Variables --------------------------------------------------------*/
 /* Exported Functions Prototypes ---------------------------------------------*/
 void MotionGC_manager_init(float freq);
-void MotionGC_manager_update(MGC_input_t *data_in, MGC_output_t *gyro_bias, int *bias_update);
+void MotionGC_manager_update(MGC_input_t *data_in, MGC_output_t *gyro_bias, int16_t *bias_update);
 void MotionGC_manager_get_knobs(MGC_knobs_t *knobs);
 void MotionGC_manager_set_knobs(MGC_knobs_t *knobs);
 void MotionGC_manager_get_params(MGC_output_t *gyro_bias);
 void MotionGC_manager_set_params(MGC_output_t *gyro_bias);
 void MotionGC_manager_set_frequency(float freq);
-void MotionGC_manager_get_version(char *version, int *length);
+void MotionGC_manager_get_version(char *version, int32_t *length);
 void MotionGC_manager_compensate(MOTION_SENSOR_Axes_t *DataIn, MOTION_SENSOR_Axes_t *DataOut);
 
 int16_t gyro_bias_to_mdps(float gyro_bias);
