+++ Projects/NUCLEO-U575ZI-Q/Examples/IKS01A3/DataLogTerminal/Src/app_mems.c
@@ -2,11 +2,11 @@
   ******************************************************************************
   * File Name          : app_mems.c
   * Description        : This file provides code for the configuration
-  *                      of the STMicroelectronics.X-CUBE-MEMS1.10.0.0 instances.
+  *                      of the STMicroelectronics.X-CUBE-MEMS1.11.0.0 instances.
   ******************************************************************************
   * @attention
   *
-  * Copyright (c) 2023 STMicroelectronics.
+  * Copyright (c) 2024 STMicroelectronics.
   * All rights reserved.
   *
   * This software is licensed under terms that can be found in the LICENSE file
@@ -100,7 +100,7 @@ void MX_MEMS_Process(void)
 void MX_IKS01A3_DataLogTerminal_Init(void)
 {
   displayFloatToInt_t out_value_odr;
-  int i;
+  int16_t i;
 
   /* Initialize LED */
   BSP_LED_Init(LED2);
@@ -184,7 +184,7 @@ void BSP_PB_Callback(Button_TypeDef Button)
   */
 void MX_IKS01A3_DataLogTerminal_Process(void)
 {
-  int i;
+  int32_t i;
 
   if (PushButtonDetected != 0U)
   {
@@ -192,7 +192,7 @@ void MX_IKS01A3_DataLogTerminal_Process(void)
     HAL_Delay(50);
 
     /* Wait until the button is released */
-    while ((BSP_PB_GetState( BUTTON_KEY ) == PushButtonState));
+    while ((BSP_PB_GetState(BUTTON_KEY) == PushButtonState));
 
     /* Debouncing */
     HAL_Delay(50);
